Refactor: feedback 컴포넌트 분리 #2-04
